many: move BuildKernelBootInfo to overlord/install package
