+++ overlord/devicestate/devicemgr.go
@@ -37,7 +37,6 @@ import (
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/gadget"
 	"github.com/snapcore/snapd/gadget/device"
-	gadgetInstall "github.com/snapcore/snapd/gadget/install"
 	"github.com/snapcore/snapd/i18n"
 	"github.com/snapcore/snapd/kernel/fde"
 	"github.com/snapcore/snapd/logger"
@@ -2249,7 +2248,7 @@ type systemAndEssentialSnaps struct {
 	*System
 	Seed            seed.Seed
 	InfosByType     map[snap.Type]*snap.Info
-	CompsByType     map[snap.Type][]gadgetInstall.CompSeedInfo
+	CompsByType     map[snap.Type][]install.CompSeedInfo
 	SeedSnapsByType map[snap.Type]*seed.Snap
 }
 
@@ -2306,7 +2305,7 @@ func (m *DeviceManager) loadSystemAndEssentialSnaps(wantedSystemLabel string, ty
 	// like "snapd" will be skipped and not part of the EssentialSnaps list
 	//
 	snapInfos := make(map[snap.Type]*snap.Info)
-	compInfos := make(map[snap.Type][]gadgetInstall.CompSeedInfo)
+	compInfos := make(map[snap.Type][]install.CompSeedInfo)
 	seedSnaps := make(map[snap.Type]*seed.Snap)
 	for _, seedSnap := range s.EssentialSnaps() {
 		typ := seedSnap.EssentialType
@@ -2330,9 +2329,9 @@ func (m *DeviceManager) loadSystemAndEssentialSnaps(wantedSystemLabel string, ty
 			return nil, fmt.Errorf("internal error while retrieving %s for %s mode: %v",
 				seedSnap.SnapName(), modeForComps, err)
 		}
-		var compInfosForType []gadgetInstall.CompSeedInfo
+		var compInfosForType []install.CompSeedInfo
 		if len(snapForMode.Components) > 0 {
-			compInfosForType = make([]gadgetInstall.CompSeedInfo, 0, len(snapForMode.Components))
+			compInfosForType = make([]install.CompSeedInfo, 0, len(snapForMode.Components))
 			for _, sc := range snapForMode.Components {
 				seedComp := sc
 				compf, err := snapfile.Open(seedComp.Path)
@@ -2344,7 +2343,7 @@ func (m *DeviceManager) loadSystemAndEssentialSnaps(wantedSystemLabel string, ty
 				if err != nil {
 					return nil, err
 				}
-				compInfosForType = append(compInfosForType, gadgetInstall.CompSeedInfo{
+				compInfosForType = append(compInfosForType, install.CompSeedInfo{
 					CompInfo: compInfo,
 					CompSeed: &seedComp,
 				})
