+++ overlord/install/install.go
@@ -38,6 +38,7 @@ import (
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/gadget"
 	"github.com/snapcore/snapd/gadget/device"
+	gadgetInstall "github.com/snapcore/snapd/gadget/install"
 	"github.com/snapcore/snapd/kernel/fde"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
@@ -81,12 +82,61 @@ type EncryptionSupportInfo struct {
 	UnavailableWarning string
 }
 
+// CompSeedInfo contains information for a component from the seed and
+// from its metadata.
+type CompSeedInfo struct {
+	CompInfo *snap.ComponentInfo
+	CompSeed *seed.Component
+}
+
+// KernelBootInfo contains information related to the kernel used on installation.
+type KernelBootInfo struct {
+	KSnapInfo     *gadgetInstall.KernelSnapInfo
+	BootableKMods []boot.BootableKModsComponents
+}
+
 var (
 	timeNow = time.Now
 
 	secbootCheckTPMKeySealingSupported = secboot.CheckTPMKeySealingSupported
 	sysconfigConfigureTargetSystem     = sysconfig.ConfigureTargetSystem
 )
+
+// BuildKernelBootInfo constructs a KernelBootInfo.
+func BuildKernelBootInfo(kernInfo *snap.Info, compSeedInfos []CompSeedInfo, kernMntPoint string, mntPtForComps map[string]string, isCore, needsDriversTree bool) KernelBootInfo {
+	bootKMods := make([]boot.BootableKModsComponents, 0, len(compSeedInfos))
+	modulesComps := make([]gadgetInstall.KernelModulesComponentInfo, 0, len(compSeedInfos))
+	for _, compSeedInfo := range compSeedInfos {
+		ci := compSeedInfo.CompInfo
+		if ci.Type == snap.KernelModulesComponent {
+			cpi := snap.MinimalComponentContainerPlaceInfo(ci.Component.ComponentName,
+				ci.Revision, kernInfo.SnapName())
+			modulesComps = append(modulesComps, gadgetInstall.KernelModulesComponentInfo{
+				Name:       ci.Component.ComponentName,
+				Revision:   ci.Revision,
+				MountPoint: mntPtForComps[ci.FullName()],
+			})
+			bootKMods = append(bootKMods, boot.BootableKModsComponents{
+				CompPlaceInfo: cpi,
+				CompPath:      compSeedInfo.CompSeed.Path,
+			})
+		}
+	}
+
+	kSnapInfo := &gadgetInstall.KernelSnapInfo{
+		Name:             kernInfo.SnapName(),
+		Revision:         kernInfo.Revision,
+		MountPoint:       kernMntPoint,
+		IsCore:           isCore,
+		ModulesComps:     modulesComps,
+		NeedsDriversTree: needsDriversTree,
+	}
+
+	return KernelBootInfo{
+		KSnapInfo:     kSnapInfo,
+		BootableKMods: bootKMods,
+	}
+}
 
 // MockSecbootCheckTPMKeySealingSupported mocks secboot.CheckTPMKeySealingSupported usage by the package for testing.
 func MockSecbootCheckTPMKeySealingSupported(f func(tpmMode secboot.TPMProvisionMode) error) (restore func()) {
